# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/summary": {
      "get": {
        "operationId": "AnalyticsController_summary_v1",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive start of the window (ISO 8601 date or datetime)",
            "schema": {
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Exclusive end of the window (ISO 8601 date or datetime)",
            "schema": {
              "example": "2026-02-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated overview for the consumer."
          }
        },
        "summary": "Overview metrics: totals, settled volume, webhook health, 30-day series",
        "tags": [
          "analytics"
        ]
      }
    },
    "/v1/balances": {
      "get": {
        "operationId": "AnalyticsController_balances_v1",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive start of the window (ISO 8601 date or datetime)",
            "schema": {
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Exclusive end of the window (ISO 8601 date or datetime)",
            "schema": {
              "example": "2026-02-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Balances per asset for the consumer."
          }
        },
        "summary": "Settled (and pending) amount per asset",
        "tags": [
          "analytics"
        ]
      }
    },
    "/v1/logs": {
      "get": {
        "operationId": "AnalyticsController_apiLogs_v1",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 500,
              "default": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API request log for the consumer."
          }
        },
        "summary": "Recent API requests reaching the service (with details)",
        "tags": [
          "analytics"
        ]
      }
    },
    "/v1/logs/webhooks": {
      "get": {
        "operationId": "AnalyticsController_webhookLogs_v1",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 500,
              "default": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook delivery log for the consumer."
          }
        },
        "summary": "Recent webhook deliveries across all endpoints (with details)",
        "tags": [
          "analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cosmos Pay — Payments API",
    "description": "Payments microservice (Stellar payment intents). All endpoints require traffic to arrive through the APISIX gateway: a valid `X-Gateway-Secret` header plus an authenticated consumer (`X-Consumer-Username`). Paths already include the version (`/v1/...`).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "gateway-secret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Gateway-Secret"
      },
      "consumer": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Consumer-Username"
      }
    },
...
  },
  "security": [
    {
      "gateway-secret": []
    },
    {
      "consumer": []
    }
  ]
}